packages/valthera/pyproject.toml
[tool.poetry]
name = "valthera"
version = "0.1.0"
description = "Universal behavioral cloning library for learning from expert demonstrations across multiple domains"
authors = ["Vijay Selvaraj <vijay.selvaraj@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/valthera/valthera"
repository = "https://github.com/valthera/valthera"
documentation = "https://valthera.readthedocs.io"
keywords = ["behavioral-cloning", "imitation-learning", "robotics", "finance", "gaming", "ai", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "valthera", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
numpy = "^1.24.0"
pandas = "^2.0.0"
pyyaml = "^6.0"
click = "^8.0.0"
tqdm = "^4.65.0"
torch = "^2.8.0"
torchvision = "^0.23.0"
torchaudio = "^2.8.0"
opencv-python = "<4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"

[tool.poetry.scripts]
valthera = "valthera.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=src/valthera",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "gpu: Tests that require GPU",
    "robotics: Robotics domain tests",
    "finance: Finance domain tests",
    "gaming: Gaming domain tests"
]
